fix loc gia tien

+++ backend/src/main/java/com/hotel/modules/room/controller/RoomController.java
@@ -59,8 +59,8 @@ public class RoomController {
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate checkIn,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate checkOut,
             @RequestParam(required = false) String province,
-            @RequestParam(required = false) Double minPrice,
-            @RequestParam(required = false) Double maxPrice,
+            @RequestParam(required = false) BigDecimal minPrice,
+            @RequestParam(required = false) BigDecimal maxPrice,
             @RequestParam(required = false) String typeName,
             @RequestParam(required = false) String bedType) {
         return ResponseEntity.ok(roomService.getAvailableRooms(
